Rename baseclass, add xmldoc & change access to internal
